Better routes for API and added back old route with new fallback
