+++ src/routes/api/jira/api-jira-uninstall-post.ts
@@ -1,0 +1,33 @@
+import { Request, Response } from "express";
+import { Installation } from "../../../models";
+import JiraClient from "../../../models/jira-client";
+import uninstall from "../../../jira/uninstall";
+
+export const ApiJiraUninstallPost = async (request: Request, response: Response): Promise<void> => {
+	response.locals.installation = await Installation.findOne({
+		where: { clientKey: request.params.clientKey }
+	});
+
+	if (!response.locals.installation) {
+		response.sendStatus(404);
+		return;
+	}
+	const jiraClient = new JiraClient(
+		response.locals.installation,
+		request.log
+	);
+	const checkAuthorization = request.body.force !== "true";
+
+	if (checkAuthorization && (await jiraClient.isAuthorized())) {
+		response
+			.status(400)
+			.json({
+				message: "Refusing to uninstall authorized Jira installation"
+			});
+		return;
+	}
+	request.log.info(
+		`Forcing uninstall for ${response.locals.installation.clientKey}`
+	);
+	await uninstall(request, response);
+};
